requirements.txt
pymupdf>=1.24.9
pydantic>=2.7
python-dateutil>=2.9
rapidfuzz>=3.9
pandas>=2.2
pytest>=8.2
fastapi>=0.115
uvicorn[standard]>=0.30
google-genai>=0.3.0
loguru>=0.7.2
openpyxl>=3.1.2
